requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
python-dotenv
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-multipart
